clickTextContaining, close #253
