+++ lib/browser_factory.js
@@ -11,7 +11,8 @@ const mixins = [
     require("./mixins/browser_info"),
     require("./mixins/browser_navigation"),
     require("./mixins/browser_queries"),
-    require("./mixins/browser_wait")
+    require("./mixins/browser_wait"),
+    require("./mixins/browser_click")
 ];
 
 module.exports = class BrowserFactory {
