+++ tests/dummy_server/static/click.html
@@ -10,6 +10,7 @@
     <button class="btn" onclick="switchText()">click me</button>
     <button class="btn2" onclick="switchTextDelay()">click me delay</button>
     <a class="btn3" onclick="switchTextDelay()">click me delay</a>
+    <a class="btn4" onclick="switchText()">weird text button</a>
     <p id="switch" class="on">On</p>
 
     <script>
